requirements.txt
influxdb==5.3.1
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic_core==2.10.1
redisai==1.3.0
websocket-client @ file:///Users/ktietz/ci_310/websocket-client_1643972661291/work
websockets==12.0
influxdb_client==1.38.0